+++ src/components/os/DesktopLayout.tsx
@@ -9,10 +9,8 @@ import { MentOS } from './MentOS';
 import { EmailApp } from '../apps/email/EmailApp';
 import { PMISApp } from '../apps/pmis/PMISApp';
 import { ChatterApp } from '../apps/chatter/ChatterApp';
-import { FilesApp } from '../apps/files/FilesApp';
 import { WikiBOKApp } from '../apps/wikibok/WikiBOKApp';
 import { ExamSimApp } from '../apps/exam/ExamSimApp';
-import ProcessMapApp from '../apps/processmap/ProcessMapApp';
 import {
   openWindow,
   closeWindow,
@@ -27,7 +25,6 @@ import {
   Globe,
   MessageCircle,
   BookOpen,
-  GitBranch,
 } from 'lucide-react';
 
 export const DesktopLayout: React.FC = () => {
@@ -60,7 +57,6 @@ export const DesktopLayout: React.FC = () => {
       case 'PMIS':
         return <LayoutDashboard size={16} />;
       case 'Document':
-      case 'Files':
         return <Folder size={16} />;
       case 'Browser':
         return <Globe size={16} />;
@@ -68,8 +64,6 @@ export const DesktopLayout: React.FC = () => {
         return <MessageCircle size={16} />;
       case 'WikiBOK':
         return <BookOpen size={16} />;
-      case 'ProcessMap':
-        return <GitBranch size={16} />;
       default:
         return <Folder size={16} />;
     }
@@ -83,12 +77,8 @@ export const DesktopLayout: React.FC = () => {
         return <PMISApp />;
       case 'Chatter':
         return <ChatterApp />;
-      case 'Files':
-        return <FilesApp />;
       case 'WikiBOK':
         return <WikiBOKApp />;
-      case 'ProcessMap':
-        return <ProcessMapApp />;
       case 'Browser':
         return <div className="p-4 text-gray-400">Browser Content Placeholder</div>;
       default: