+++ src/store.ts
@@ -5,7 +5,6 @@ import pmisReducer from './features/pmisSlice';
 import osReducer from './features/osSlice';
 import dialogueReducer from './features/dialogueSlice';
 import examReducer from './features/examSlice';
-import processMapReducer from './features/processMapSlice';
 import documentAnalysisReducer from './features/documentAnalysisSlice';
 import pmisEvolutionReducer from './features/pmisEvolutionSlice';
 import decisionReducer from './features/decisionSlice';
@@ -19,7 +18,6 @@ export const store = configureStore({
     os: osReducer,
     dialogue: dialogueReducer,
     exam: examReducer,
-    processMap: processMapReducer,
     documentAnalysis: documentAnalysisReducer,
     pmisEvolution: pmisEvolutionReducer,
     decisions: decisionReducer,